### 리팩토링
1. 디렉터리 생성 및 파일 정리
2. Preset 값 enum namespace 로 정의
3. 불필요한 선언 제거
4. RunGuideTabView 서브뷰 추출 및 정리
